+++ uiza/__init__.py
@@ -1,4 +1,5 @@
 from uiza.api_resources import *
 
-workspace_api_domain = None
+workspace_api_domain = 'stag-ap-southeast-1-api.uizadev.io'
 api_key = None
+app_id = None