+++ uiza/api_resources/user/__init__.py
@@ -1,1 +1,0 @@
-from uiza.api_resources.user.user import User